requirements.txt
asgiref==3.2.10
boto3==1.15.16
botocore==1.18.16
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.1
django-allauth==0.42.0
django-bootstrap-datepicker==1.4.0
django-countries==6.1.3
django-crispy-forms==1.9.2
django-csvimport==2.16
django-import-export==2.4.0
django-storages==1.10.1
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
jmespath==0.10.0
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
Pillow==7.2.0
postcode.io==0.1.1
psycopg2-binary==2.8.6
python3-openid==3.2.0
pytz==2020.1
requests-oauthlib==1.3.0
s3transfer==0.3.3
sqlparse==0.3.1
stripe==2.53.0
tablib==2.0.0
xlrd==1.2.0
xlwt==1.3.0
